es-python-flask/requirements.txt
Flask==2.0.1
flask-restful==0.3.9
webargs==8.0.0
marshmallow==3.13.0
flask-apispec==0.11.0
elasticsearch-dsl==7.4.0
click==7.1.2
python-dotenv==0.19.0
pyyaml
rootpath==0.1.1
flake8==3.9.2
pytest==6.2.4
mock==4.0.2